Docker/python/requirements.txt
pyserial
psycopg2
